Revert "base: update settings material colors to match systemUI"
This reverts commit 4325164c0b61ac402c917225c3b91baea5832ea1.

I think it's better, when we use the Stock Android Primary Colors
instead of the Pixel Primary because right now,
we have Primary Pixel and Stock Accent Colors.

These Stock Colors were been used in CM12.1 and also in CM13,
why not in CM14.1. It looks more clean with these Colors.
Many Rom’s switched to the Pixel look,
LineageOS is one of the last Custom Rom’s with Android 7.1.1
and an Half-Stock look. The Stock look would be Perfect for our Rom.
It’s not a Big Change but a Good Change.

[LineageOS right now: http://i.imgur.com/8Cujj0q.png]
[LineageOS after this Color Change: http://i.imgur.com/M7wA20X.png]

Change-Id: Ie214eddfc2f1ae926bf09907949c6fb94b82c031